refactor: rename modules to appropriate names
- services/payment.py -> payment_service.py
- schemas/payment.py -> payment_schemas.py
- star_price.py -> star_service.py
