+++ core/services/star_service.py
